+++ lib/src/helpers/default_list_tile.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:simple_search_dropdown/simple_search_dropdown.dart';
 
 ///The default widget that will be presented on the list.
-class DefaultListTile<T> extends StatelessWidget {
+class DefaultListTile<T> extends StatefulWidget {
   const DefaultListTile({
     required this.deleteMode,
     required this.editMode,
@@ -28,9 +28,38 @@ class DefaultListTile<T> extends StatelessWidget {
   final Widget? defaultAditionalWidget;
 
   @override
+  State<DefaultListTile<T>> createState() => _DefaultListTileState<T>();
+}
+
+class _DefaultListTileState<T> extends State<DefaultListTile<T>> {
+  bool _isLoading = false;
+
+  Future<void> _handleDeleteItem() async {
+    setState(() => _isLoading = true);
+    try {
+      await widget.onDelete(widget.item);
+    } finally {
+      if (mounted) {
+        setState(() => _isLoading = false);
+      }
+    }
+  }
+
+  Future<void> _handleEditItem() async {
+    setState(() => _isLoading = true);
+    try {
+      await widget.onEdit(widget.item);
+    } finally {
+      if (mounted) {
+        setState(() => _isLoading = false);
+      }
+    }
+  }
+
+  @override
   Widget build(BuildContext context) {
     return Padding(
-      padding: overlayListSettings.itemsPadding,
+      padding: widget.overlayListSettings.itemsPadding,
       child: Row(
         children: [
           Expanded(
@@ -38,8 +67,9 @@ class DefaultListTile<T> extends StatelessWidget {
               style: ButtonStyle(
                 backgroundColor: WidgetStateProperty.resolveWith<Color>(
                   (Set<WidgetState> states) {
-                    if (selected) {
-                      return overlayListSettings.selectedItemBackgroundColor;
+                    if (widget.selected) {
+                      return widget
+                          .overlayListSettings.selectedItemBackgroundColor;
                     }
                     return Colors.transparent;
                   },
@@ -51,50 +81,57 @@ class DefaultListTile<T> extends StatelessWidget {
                 ),
                 overlayColor: WidgetStateProperty.resolveWith<Color>(
                   (Set<WidgetState> states) {
-                    if (selected) {
-                      return overlayListSettings.selectedItemHoverColor;
+                    if (widget.selected) {
+                      return widget.overlayListSettings.selectedItemHoverColor;
                     }
-                    return overlayListSettings.unselectedItemHoverColor;
+                    return widget.overlayListSettings.unselectedItemHoverColor;
                   },
                 ),
               ),
-              onPressed: () => onPressed(item),
-              child: overlayListSettings.itemWidgetBuilder != null
-                  ? overlayListSettings.itemWidgetBuilder!(item)
+              onPressed: () => widget.onPressed(widget.item),
+              child: widget.overlayListSettings.itemWidgetBuilder != null
+                  ? widget.overlayListSettings.itemWidgetBuilder!(widget.item)
                   : Align(
                       alignment: Alignment.centerLeft,
                       child: Text(
-                        item.label,
+                        widget.item.label,
                         maxLines: 2,
                         overflow: TextOverflow.ellipsis,
-                        style: (selected
-                            ? overlayListSettings.selectedItemTextStyle
-                            : overlayListSettings.unselectedItemTextStyle),
+                        style: (widget.selected
+                            ? widget.overlayListSettings.selectedItemTextStyle
+                            : widget
+                                .overlayListSettings.unselectedItemTextStyle),
                       ),
                     ),
             ),
           ),
-          if (defaultAditionalWidget != null) ...[
-            defaultAditionalWidget!,
-            SizedBox(width: overlayListSettings.aditionalWidgetSpacing)
+          if (widget.defaultAditionalWidget != null) ...[
+            widget.defaultAditionalWidget!,
+            SizedBox(width: widget.overlayListSettings.aditionalWidgetSpacing)
           ],
-          if (editMode)
-            IconButton(
-              onPressed: () {
-                onEdit(item);
-              },
-              icon: overlayListSettings.dialogEditIcon,
-            ),
-          if (deleteMode)
-            overlayListSettings.dialogActionWidget ??
-                IconButton(
-                  onPressed: () {
-                    onDelete(item);
-                  },
-                  icon: overlayListSettings.dialogDeleteIcon,
-                )
-          else
-            const SizedBox.shrink()
+          if (_isLoading)
+            Padding(
+              padding: const EdgeInsets.only(left: 6,right: 6),
+              child: SizedBox(
+                height: widget.overlayListSettings.loadingSize,
+                child: Loading(cor: widget.overlayListSettings.loadingColor),
+              ),
+            )
+          else ...[
+            if (widget.editMode)
+              IconButton(
+                onPressed: _handleEditItem,
+                icon: widget.overlayListSettings.dialogEditIcon,
+              ),
+            if (widget.deleteMode)
+              widget.overlayListSettings.dialogActionWidget ??
+                  IconButton(
+                    onPressed: _handleDeleteItem,
+                    icon: widget.overlayListSettings.dialogDeleteIcon,
+                  )
+            else
+              const SizedBox.shrink()
+          ],
         ],
       ),
     );