+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [7.7.0] - 2026-01-27.
+- Added also the loading widget to the delete button on the list cell. 
+- Removed `onClear` function from the controller. You can get the similar result by using `updateSelectedItem`.
+
 ## [7.6.0] - 2026-01-27.
 - Added a loading widget to the list cell. 
 - Added `loadingColor` and `loadingSize` parameters to `SimpleOverlaySettings`.